refactor(admin): AdminUsers sous 300 lignes — éclatement complet (#G1 ≤300)
AdminUsers.vue : 819 -> 206 lignes (fichier entier ≤300). Décomposition :
- composables/useAdminUsers.js (147) : état + agrégation étudiants/enseignants,
  filtres/tri/pagination, cache + refresh arrière-plan.
- components/admin/UsersStatsCards.vue (75) : cartes de stats.
- components/admin/UsersFilters.vue (117) : recherche + rôle + classe (v-model).
- components/admin/UsersTable.vue (242) : table triable + pagination (v-model).
La vue devient un orchestrateur ; le CSS est distribué dans chaque sous-composant.

Tests : useAdminUsers (agrégation, filtres, tri, pagination — services mockés),
UsersStatsCards, UsersFilters, UsersTable. 16/16 verts. vite build vert.
